Address review comments

+++ db/migrate/20161105080056_create_agency_template_uploads.rb
@@ -3,9 +3,9 @@ class CreateAgencyTemplateUploads < ActiveRecord::Migration
     create_table :agency_template_uploads do |t|
       t.timestamp :uploaded_at
       t.references :scheme, index: true, foreign_key: true
-      t.float :uploaded_by_id
+      t.integer :uploaded_by_id
       t.string :uploaded_by_type
-      t.float :year
+      t.integer :year
       t.string :status
       t.string :filename
