+++ db/migrate/20161204142200_create_leavers.rb
@@ -5,7 +5,7 @@ class CreateLeavers < ActiveRecord::Migration
       t.references :leaving_code, foreign_key: true
       t.references :agency_template_upload, foreign_key: true
       t.date :date, null: false
-      t.float :total_recovery_previous
+      t.decimal :total_recovery_previous
       t.boolean :sub_leaver
       t.date :scheme_registration_date
