+++ db/migrate/20161204153226_create_joiners.rb
@@ -5,7 +5,7 @@ class CreateJoiners < ActiveRecord::Migration
       t.references :business, foreign_key: true
       t.date :joining_date
       t.string :previously_registered_at
-      t.float :total_recovery
+      t.decimal :total_recovery
 
       t.timestamps null: false
     end
