+++ db/migrate/20161204211948_create_annual_target_sets.rb
@@ -2,7 +2,7 @@ class CreateAnnualTargetSets < ActiveRecord::Migration
   def change
     create_table :annual_target_sets do |t|
       t.references :scheme_country_code, foreign_key: true
-      t.float :value
+      t.decimal :value
       t.string :year
 
       t.timestamps null: false
