+++ db/migrate/20161204211953_create_target_totals.rb
@@ -2,11 +2,11 @@ class CreateTargetTotals < ActiveRecord::Migration
   def change
     create_table :target_totals do |t|
       t.references :regular_producer_detail, foreign_key: true
-      t.float :total_recycling_obligation
-      t.float :total_recovery_obligation
-      t.float :total_material_specific_recycling_obligation
-      t.float :adjusted_total_recovery_obligation
-      t.float :ninetytwo_percent_min_recycling_target
+      t.decimal :total_recycling_obligation
+      t.decimal :total_recovery_obligation
+      t.decimal :total_material_specific_recycling_obligation
+      t.decimal :adjusted_total_recovery_obligation
+      t.decimal :ninetytwo_percent_min_recycling_target
 
       t.timestamps null: false
     end
