+++ db/migrate/20161204211955_create_material_details.rb
@@ -4,23 +4,23 @@ class CreateMaterialDetails < ActiveRecord::Migration
       t.references :regular_producer_detail, foreign_key: true
       t.references :packaging_material, foreign_key: true
 
-      t.float :t1manwo
-      t.float :t1conv
-      t.float :t1pf
-      t.float :t1sell
-      t.float :t2aman
-      t.float :t2conv
-      t.float :t2apf
-      t.float :t2sell
-      t.float :t2bman
-      t.float :t2bconv
-      t.float :t2bp
-      t.float :t2bsell
-      t.float :t3aconv
-      t.float :t3apf
-      t.float :t3asell
-      t.float :t3b
-      t.float :t3c
+      t.decimal :t1manwo
+      t.decimal :t1conv
+      t.decimal :t1pf
+      t.decimal :t1sell
+      t.decimal :t2aman
+      t.decimal :t2conv
+      t.decimal :t2apf
+      t.decimal :t2sell
+      t.decimal :t2bman
+      t.decimal :t2bconv
+      t.decimal :t2bp
+      t.decimal :t2bsell
+      t.decimal :t3aconv
+      t.decimal :t3apf
+      t.decimal :t3asell
+      t.decimal :t3b
+      t.decimal :t3c
 
       t.timestamps null: false
     end
