+++ db/migrate/20161204211956_create_material_targets.rb
@@ -4,7 +4,7 @@ class CreateMaterialTargets < ActiveRecord::Migration
       t.references :packaging_material, foreign_key: true
       t.references :annual_target_set, foreign_key: true
       t.string :year
-      t.float :value
+      t.decimal :value
 
       t.timestamps null: false
     end
